requirements.txt
Pillow>=10.0.0
crc8>=0.2.0
pexpect>=4.9
